Nested Sortable Update
Added in a quite awful example of a nested sortable list. Above the initializer there are comments on how to group them, though I ran out of time to get it to work helping drivers with their tablets. So, follow the comments it should be straight forward. Also there is a way to save the order preference to a DB, just gotta figure out what values to store. Maybe create a sortId?